fix: update bulk user without id
